requirements.txt
api-client==1.3.1
asgiref==3.8.1
asttokens==2.4.1
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==43.0.0
decorator==5.1.1
Django==5.1
django-rest-swagger==2.2.0
djangorestframework==3.15.2
drf-yasg==1.21.7
executing==2.0.1
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.26.0
itypes==1.2.0
jedi==0.19.1
Jinja2==3.1.4
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
openapi-codec==1.3.2
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
pluggy==1.5.0
prompt_toolkit==3.0.47
psycopg==3.2.1
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
pytest==8.3.2
pytest-django==4.8.0
pytest-env==1.1.3
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
simplejson==3.19.3
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
svgwrite==1.4.3
tenacity==9.0.0
traitlets==5.14.3
Tree==0.2.4
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
wcwidth==0.2.13
